# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * LearnWise AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  displayName: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  studyGoal?: string | null;
  /** @nullable */
  weeklyGoalHours?: number | null;
  createdAt: string;
}

export interface UserProfileUpdate {
  displayName?: string;
  bio?: string;
  studyGoal?: string;
  weeklyGoalHours?: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  quiz: 'quiz',
  flashcard: 'flashcard',
  pdf: 'pdf',
  study_plan: 'study_plan',
  tutor: 'tutor',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  description: string;
  createdAt: string;
}

export interface ProgressSummary {
  totalStudyMinutes: number;
  currentStreak: number;
  weeklyGoalHours: number;
  weeklyMinutesStudied: number;
  quizzesCompleted: number;
  flashcardsReviewed: number;
  pdfsUploaded: number;
  studyPlansActive: number;
}

export type QuizStatus = typeof QuizStatus[keyof typeof QuizStatus];


export const QuizStatus = {
  pending: 'pending',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface Quiz {
  id: number;
  userId: number;
  title: string;
  topic: string;
  questionCount: number;
  /** @nullable */
  score: number | null;
  status: QuizStatus;
  createdAt: string;
}

export interface QuizInput {
  title: string;
  topic: string;
  /**
     * @minimum 1
     * @maximum 50
     */
  questionCount: number;
}

export interface FlashcardSet {
  id: number;
  userId: number;
  title: string;
  topic: string;
  cardCount: number;
  /** @nullable */
  lastStudiedAt?: string | null;
  createdAt: string;
}

export interface FlashcardSetInput {
  title: string;
  topic: string;
  /** @minimum 1 */
  cardCount?: number;
}

export type StudyPlanStatus = typeof StudyPlanStatus[keyof typeof StudyPlanStatus];


export const StudyPlanStatus = {
  active: 'active',
  completed: 'completed',
  paused: 'paused',
} as const;

export interface StudyPlan {
  id: number;
  userId: number;
  title: string;
  subject: string;
  durationWeeks: number;
  hoursPerWeek: number;
  status: StudyPlanStatus;
  progressPercent?: number;
  createdAt: string;
}

export interface StudyPlanInput {
  title: string;
  subject: string;
  /** @minimum 1 */
  durationWeeks: number;
  /** @minimum 1 */
  hoursPerWeek: number;
}

export type StudyPlanUpdateStatus = typeof StudyPlanUpdateStatus[keyof typeof StudyPlanUpdateStatus];


export const StudyPlanUpdateStatus = {
  active: 'active',
  completed: 'completed',
  paused: 'paused',
} as const;

export interface StudyPlanUpdate {
  title?: string;
  subject?: string;
  durationWeeks?: number;
  hoursPerWeek?: number;
  status?: StudyPlanUpdateStatus;
  progressPercent?: number;
}

export type PdfDocumentStatus = typeof PdfDocumentStatus[keyof typeof PdfDocumentStatus];


export const PdfDocumentStatus = {
  processing: 'processing',
  ready: 'ready',
  error: 'error',
} as const;

export interface PdfDocument {
  id: number;
  userId: number;
  fileName: string;
  originalName: string;
  fileSize: number;
  /** @nullable */
  pageCount?: number | null;
  status: PdfDocumentStatus;
  createdAt: string;
}

export interface PdfInput {
  fileName: string;
  originalName: string;
  fileSize: number;
  pageCount?: number;
}

export type TutorSessionStatus = typeof TutorSessionStatus[keyof typeof TutorSessionStatus];


export const TutorSessionStatus = {
  active: 'active',
  ended: 'ended',
} as const;

export interface TutorSession {
  id: number;
  userId: number;
  subject: string;
  messageCount: number;
  status: TutorSessionStatus;
  /** @nullable */
  lastMessageAt?: string | null;
  createdAt: string;
}

export interface TutorSessionInput {
  subject: string;
}