# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LearnWise AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quizzes:
    get:
      operationId: getQuizzes
      tags: [quizzes]
      summary: List quiz history
      responses:
        "200":
          description: List of quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create a new quiz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizInput"
      responses:
        "201":
          description: Created quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get quiz by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
        "404":
          description: Not found
    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      summary: Delete a quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Flashcards ─────────────────────────────────────────────────────────────
  /flashcards:
    get:
      operationId: getFlashcardSets
      tags: [flashcards]
      summary: List flashcard sets
      responses:
        "200":
          description: List of flashcard sets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlashcardSet"
    post:
      operationId: createFlashcardSet
      tags: [flashcards]
      summary: Create a flashcard set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardSetInput"
      responses:
        "201":
          description: Created flashcard set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardSet"
  /flashcards/{id}:
    get:
      operationId: getFlashcardSet
      tags: [flashcards]
      summary: Get a flashcard set by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Flashcard set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardSet"
        "404":
          description: Not found
    delete:
      operationId: deleteFlashcardSet
      tags: [flashcards]
      summary: Delete a flashcard set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Study Plans ─────────────────────────────────────────────────────────────
...
  /pdfs:
    get:
      operationId: getPdfs
      tags: [pdfs]
      summary: List uploaded PDFs
      responses:
        "200":
          description: List of PDFs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PdfDocument"
    post:
      operationId: createPdf
      tags: [pdfs]
      summary: Register a PDF upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PdfInput"
      responses:
        "201":
          description: Registered PDF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdfDocument"
  /pdfs/{id}:
    get:
      operationId: getPdf
      tags: [pdfs]
      summary: Get a PDF by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdfDocument"
        "404":
          description: Not found
    delete:
      operationId: deletePdf
      tags: [pdfs]
      summary: Delete a PDF record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── AI Tutor Sessions ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Quiz:
      type: object
      required: [id, userId, title, topic, questionCount, score, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        topic:
          type: string
        questionCount:
          type: integer
        score:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, completed, archived]
        createdAt:
          type: string
          format: date-time
    QuizInput:
      type: object
      required: [title, topic, questionCount]
      properties:
        title:
          type: string
        topic:
          type: string
        questionCount:
          type: integer
          minimum: 1
          maximum: 50

    # ─── Flashcards ──────────────────────────────────────────────────────────
    FlashcardSet:
      type: object
      required: [id, userId, title, topic, cardCount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        topic:
          type: string
        cardCount:
          type: integer
        lastStudiedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    FlashcardSetInput:
      type: object
      required: [title, topic]
      properties:
        title:
          type: string
        topic:
          type: string
        cardCount:
          type: integer
          minimum: 1

    # ─── Study Plans ─────────────────────────────────────────────────────────
...
    PdfDocument:
      type: object
      required: [id, userId, fileName, originalName, fileSize, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        fileName:
          type: string
        originalName:
          type: string
        fileSize:
          type: integer
        pageCount:
          type: ["integer", "null"]
        status:
          type: string
          enum: [processing, ready, error]
        createdAt:
          type: string
          format: date-time
    PdfInput:
      type: object
      required: [fileName, originalName, fileSize]
      properties:
        fileName:
          type: string
        originalName:
          type: string
        fileSize:
          type: integer
        pageCount:
          type: integer

    # ─── Tutor Sessions ──────────────────────────────────────────────────────
...