# lib/api-zod/src/generated/types/flashcardSet.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * LearnWise AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FlashcardSet {
  id: number;
  userId: number;
  title: string;
  topic: string;
  cardCount: number;
  /** @nullable */
  lastStudiedAt?: Date | null;
  createdAt: Date;
}