# lib/api-zod/src/generated/types/pdfDocument.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * LearnWise AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PdfDocumentStatus } from './pdfDocumentStatus';

export interface PdfDocument {
  id: number;
  userId: number;
  fileName: string;
  originalName: string;
  fileSize: number;
  /** @nullable */
  pageCount?: number | null;
  status: PdfDocumentStatus;
  createdAt: Date;
}