# lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * LearnWise AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizStatus } from './quizStatus';

export interface Quiz {
  id: number;
  userId: number;
  title: string;
  topic: string;
  questionCount: number;
  /** @nullable */
  score: number | null;
  status: QuizStatus;
  createdAt: Date;
}