# lib/api-zod/src/generated/types/quizInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * LearnWise AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuizInput {
  title: string;
  topic: string;
  /**
     * @minimum 1
     * @maximum 50
     */
  questionCount: number;
}